requirements.txt
openai-whisper
pyaudio
numpy
torch
tkinter 